Partially implemented inventory listing

+++ include/map.h
@@ -96,7 +96,7 @@ class Map
     void SetTile(int x, int y,tile t);
 
     Item* GetItem(int x, int y);
-    Item* GetContainerItem(int containerX, int containerY,int index);
+    Item* GetContainerItem(int containerX, int containerY,unsigned long int index);
     
     void AddToContainer(int x, int y,Item* i);
